wrap metrics summaries in newtype
