+++ relay-event-schema/src/protocol/mod.rs
@@ -16,6 +16,7 @@ mod logentry;
 mod measurements;
 mod mechanism;
 mod metrics;
+mod metrics_summary;
 mod nel;
 mod relay_info;
 mod replay;
@@ -53,6 +54,7 @@ pub use self::logentry::*;
 pub use self::measurements::*;
 pub use self::mechanism::*;
 pub use self::metrics::*;
+pub use self::metrics_summary::*;
 pub use self::nel::*;
 pub use self::relay_info::*;
 pub use self::replay::*;
