+++ relay-event-schema/src/protocol/span.rs
@@ -1,35 +1,12 @@
 #[cfg(feature = "jsonschema")]
 use relay_jsonschema_derive::JsonSchema;
-use relay_protocol::{Annotated, Array, Empty, FromValue, Getter, IntoValue, Object, Val, Value};
+use relay_protocol::{Annotated, Empty, FromValue, Getter, IntoValue, Object, Val, Value};
 
 use crate::processor::ProcessValue;
 use crate::protocol::{
-    Event, EventId, JsonLenientString, Measurements, OperationType, OriginType, ProfileContext,
-    SpanId, SpanStatus, Timestamp, TraceContext, TraceId,
+    Event, EventId, JsonLenientString, Measurements, MetricsSummary, OperationType, OriginType,
+    ProfileContext, SpanId, SpanStatus, Timestamp, TraceContext, TraceId,
 };
-
-/// The metric summary of a single metric that is emitted by the SDK.
-///
-/// The summary contains specific aggregate values that the metric had during the span's lifetime. A single span can
-/// have the same metric emitted multiple times, which is the reason for aggregates being computed in each summary.
-#[derive(Clone, Debug, Default, PartialEq, Empty, FromValue, IntoValue, ProcessValue)]
-#[cfg_attr(feature = "jsonschema", derive(JsonSchema))]
-pub struct MetricSummary {
-    /// Minimum value of the metric.
-    pub min: Annotated<f64>,
-
-    /// Maximum value of the metric.
-    pub max: Annotated<f64>,
-
-    /// Sum of all metric values.
-    pub sum: Annotated<f64>,
-
-    /// Count of all metric values.
-    pub count: Annotated<u64>,
-
-    /// Tags of the metric.
-    pub tags: Annotated<Object<String>>,
-}
 
 #[derive(Clone, Debug, Default, PartialEq, Empty, FromValue, IntoValue, ProcessValue)]
 #[cfg_attr(feature = "jsonschema", derive(JsonSchema))]
@@ -109,7 +86,7 @@ pub struct Span {
     /// This shall move to a stable location once we have stabilized the
     /// interface.  This is intentionally not typed today.
     #[metastructure(skip_serialization = "empty")]
-    pub _metrics_summary: Annotated<Object<Array<MetricSummary>>>,
+    pub _metrics_summary: Annotated<MetricsSummary>,
 
     // TODO remove retain when the api stabilizes
     /// Additional arbitrary fields for forwards compatibility.
@@ -329,19 +306,21 @@ mod tests {
             sentry_tags: ~,
             received: ~,
             measurements: ~,
-            _metrics_summary: {
-                "some_metric": [
-                    MetricSummary {
-                        min: 1.0,
-                        max: 2.0,
-                        sum: 3.0,
-                        count: 2,
-                        tags: {
-                            "environment": "test",
+            _metrics_summary: MetricsSummary(
+                {
+                    "some_metric": [
+                        MetricSummary {
+                            min: 1.0,
+                            max: 2.0,
+                            sum: 3.0,
+                            count: 2,
+                            tags: {
+                                "environment": "test",
+                            },
                         },
-                    },
-                ],
-            },
+                    ],
+                },
+            ),
             other: {},
         }
         "###);
